WIP: Prototype public transitive headermap rule
This patch prototypes a headermap rule in skylark!

It works quite well, and sacrifices usability, since it isn't implicitly
created.

Public transitive `objc_library` header usage:
- `deps` go into `hdr_providers`
- `hdrs` into `hdrs` wrapped in a filegroup
- the actual header map goes into `hdrs` for public usage

Other consumers of the rules can wrap `objc_library` in a macro, which
both instantiate and consume this rule.

It works by transitively merging public headermaps. This works
effortlessly, since headermaps are propagated transitively along with the
headers.

Usage:
Create a basic headermap ( for the pod PINOperation )
headermap(
  name = "PINOperation_hmap",
  namespace = "PINOperation",
  # File group to headers
  hdrs = [
    "PINOperation_hdrs"
  ],
  hdr_providers = []
)

It supports transitive propagation via the `hdr_providers` rule,
which are currently objc providers only, we can expand this to multiple
types once swift support is added.
headermap(
  name = "PINCache_hmap",
  namespace = "PINCache",
  hdrs = [
    "PINCache_hdrs"
  ],
  hdr_providers = [ "//Vendor/PINOperation:PINOperation" ]
)

Compilation flags:
.hmaps will need to be explicitly added to the command line. Likely,
this will be achieved with `copts` in PodToBUILD. I'm not sure if the
behavior of replacing `GENFILES` with the correct directory is
available in Bazel, so that may rely on a custom toolchain or simply use
the `include` attribute.

Other considerations:
Be smarter about headers that actually exist. This needs more thought.
If the input pruning and dot d pruning work is useable from skylark,
it'd be great to leverage that.